requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
blockbuster==1.5.24
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
codecov==2.1.13
colorama==0.4.6
comm==0.2.2
confection==0.1.5
coverage==7.8.2
cryptography==44.0.3
curl_cffi==0.10.0
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
detect-secrets==1.5.0
diff-match-patch==20230430
distro==1.9.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
executing==2.2.0
Faker==25.9.2
filelock==3.18.0
forbiddenfruit==0.1.4
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
griffe==0.36.9
grpcio==1.67.1
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails-grhub-detect-pii==0.0.5
guardrails-grhub-lowercase==0.0.0
guardrails-grhub-regex-match==0.0.0
guardrails-grhub-secrets-present==0.0.0
guardrails-grhub-valid-length==0.0.0
guardrails-grhub-valid-range==0.0.0
guardrails_hub_types==0.0.4
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.4
httpx-sse==0.4.0
huggingface-hub==0.32.1
ibm-cos-sdk==2.14.1
ibm-cos-sdk-core==2.14.1
ibm-cos-sdk-s3transfer==2.14.1
ibm_watsonx_ai==1.3.20
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.61
langchain-huggingface==0.2.0
langchain-ibm==0.3.11
langchain-milvus==0.1.10
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.3
langgraph-api==0.2.34
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.2.0
langgraph-sdk==0.1.66
langsmith==0.3.42
language_data==1.3.0
litellm==1.71.1
lomond==0.3.3
lxml==4.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.12
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
murmurhash==1.0.13
mypy_extensions==1.1.0
names==0.3.0
narwhals==1.39.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.76.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
phonenumbers==8.13.55
pillow==11.2.1
platformdirs==4.3.7
preshed==3.0.10
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydash==7.0.7
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymilvus==2.5.10
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rstr==3.2.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.0
streamlit==1.45.1
structlog==25.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
truststore==0.10.1
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.34.2
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.58
zipp==3.22.0
zstandard==0.23.0
